re-org project files
