+++ Source/nHydrate.Generator.Common/GeneratorFramework/IProjectGenerator.cs
@@ -5,11 +5,15 @@ namespace nHydrate.Generator.Common.GeneratorFramework
         void Initialize(IModelObject model);
         string ProjectName { get; }
         void CreateProject();
-        //void LoadProject();
         string DefaultNamespace { get; }
         string LocalNamespaceExtension { get; }
         string GetLocalNamespace();
         IModelConfiguration ModelConfiguration { get; set; }
+        string ProjectTemplate { get; }
+        void GenerateCompanySpecificFile(string tempPath, string fileName);
+        void OnAfterGenerate();
+        Models.ModelRoot Model { get; }
+        nHydrate.Generator.Common.ProjectItemGenerators.IProjectGeneratorProjectCreator ProjectGeneratorProjectCreator { get; set; }
     }
 
     public interface IModelConfiguration
