+++ Source/nHydrate.Generator.Common/ProjectItemGenerators/BaseClassGenerator.cs
@@ -1,0 +1,7 @@
+namespace nHydrate.Generator.Common.ProjectItemGenerators
+{
+    public abstract class BaseClassGenerator : BaseProjectItemGenerator
+    {
+    }
+}
+
