+++ Source/nHydrate.Generator.Common/nHydrate.Generator.Common.csproj
@@ -215,6 +215,10 @@
     <Compile Include="Models\VersionHistory.cs" />
     <Compile Include="Models\VersionHistoryCollection.cs" />
     <Compile Include="nHydrateGeneratorProject.cs" />
+    <Compile Include="ProjectItemGenerators\BaseClassGenerator.cs" />
+    <Compile Include="ProjectItemGenerators\BaseClassTemplate.cs" />
+    <Compile Include="ProjectItemGenerators\BaseProjectGenerator.cs" />
+    <Compile Include="ProjectItemGenerators\BaseProjectItemGenerator.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
