+++ Source/nHydrate.Generator.EFCodeFirstNetCore/EFCodeFirstNetCoreBaseTemplate.cs
@@ -1,5 +1,5 @@
 using nHydrate.Generator.Common.Models;
-using nHydrate.Generator.ProjectItemGenerators;
+using nHydrate.Generator.Common.ProjectItemGenerators;
 
 namespace nHydrate.Generator.EFCodeFirstNetCore
 {
