--- Source/nHydrate.Generator/ProjectItemGenerators/BaseClassGenerator.cs
@@ -1,7 +1,0 @@
-namespace nHydrate.Generator.ProjectItemGenerators
-{
-    public abstract class BaseClassGenerator : BaseProjectItemGenerator
-    {
-    }
-}
-
