+++ Source/nHydrate.Generator/nHydrate.Generator.csproj
@@ -219,7 +219,7 @@
     <Compile Include="ModelUI\StatisticsForm.Designer.cs">
       <DependentUpon>StatisticsForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="ProjectItemGenerators\BaseProjectGenerator.cs" />
+    <Compile Include="ProjectItemGenerators\ProjectGeneratorProjectCreator.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
@@ -230,15 +230,6 @@
       <AutoGen>True</AutoGen>
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
       <DependentUpon>Settings.settings</DependentUpon>
-    </Compile>
-    <Compile Include="ProjectItemGenerators\BaseClassGenerator.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Include="ProjectItemGenerators\BaseClassTemplate.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Include="ProjectItemGenerators\BaseProjectItemGenerator.cs">
-      <SubType>Code</SubType>
     </Compile>
     <Compile Include="Util\EnvDTEHelper.cs" />
     <Compile Include="Util\Extensions.cs" />
